Create separate functions to improve performance
Moved the event listeners to a separate function that is only called
once (not for every iteration of the createGrid function, as was
previously the case); this function creates a nodelist from all the new
grid cells, then uses the forEach method to apply the listeners to each
cell.

Also applied border-radius rounding to the cells to make them look more
like pointillist dots.